// src/app/interfaces/vehiculo.ts
export interface Vehiculo {
    id_vehiculo?: number;
    id_dispositivo?: number;
    id_master?: number;
    id_usuario?: number;
    imei?: string;
    nombre?: string;
    rxart?: string;
    interface?: string;
    imsi?: string;
    operador?: string;
    id_sim?: string;
    version?: string;
    registrado?: string;
    year_vehicle?: string;
    marca?: string;
    modelo?: string;
    placa?: string;
    color?: string;
    vin?: string;
    descripcion?: string;
    asociado?: string;
}